Attempt to obtain name of top directory in archive from archive content
The top archive directory is not always identical to the archive name.
This is for instance the case when a construct:
    Source: http://...#/foo.tar.gz
is used in the spec file.
Therefore attempt to gather it from the archive itself.
If this fails (possibly because there are multiple top directories
in the archive), fall back to the old method.
Look for go.mod below this directory.

Signed-off-by: Egbert Eich <eich@suse.com>