Add support to modify vendor module versions
The module update information is maintained in a YAML file. This service
will read the YAML file, apply the changes, create a patch and update
the changelog file. Once this is done, it will proceed creating the
vendor tarball.
The advantage if this approach over patching go.mod and go.sum directly
is that the YAML file can be created automatically without the need to
untar the source tarball. This is all done by `go_modules`.

Added functionality in detail:
Once the source tar ball is expanded, this performs the following steps:
1. Parse a YAML file with information on module versions that require
   changes,
2. apply these changes using `go get <module>[@<version>`,
3. Add any comments that do not exist yet in .changes file(s) to these.
4. create a patch that can be used with in the spec file. If no changes
   are required, the patch file will be empty.
For this 3 new arguments are introduced:
* `modupdates`: name of the YAML file. Default is 'none'.
* `moddiff`: name of the patch file to generate. The patch file will
   be written to `outdir`. Default: `go-modules.patch`
* 'changesfile': if multiple changes files exist. by default all of
  them are checked for missing comments and comments are added to the
  ones where it is missing. This option allows to specify an individual
  changes file.
The YAML file has the following format:
go_module:
   module_updates:
   - uri: <module>@<version>
     comment: <comment>
the uri list elements are passed to `go get` verbatim. A minimal check for
unsupported characters is performed. The `comment` element is optional. Any
comment is checked against the changes file(s) if it exist already. If not,
a new comment is created with all comment elements that don't already exist.
Multi-line comments are possible, line-breaks need to be indicated by `\n`.

Signed-off-by: Egbert Eich <eich@suse.com>